Add ExpandableServiceItemRow component to handle expanded/collapsed state
