+++ src/components/Office/PaymentRequestDetails/PaymentRequestDetails.module.scss
@@ -76,19 +76,23 @@
     }
   }
 
-  .expandable {
+  .isExpandable {
     cursor: pointer;
+  }
+
+  .expandable {
     .icon {
       @include u-margin-left(-1);
       @include u-margin-right(2);
       color: $primary;
     }
 
-    &.expandedRow, &.expandedDetail {
+    &.expandedRow,
+    &.expandedDetail {
       border: 1px solid $base-lighter;
     }
 
-    &.expandedRow td{
+    &.expandedRow td {
       background-color: $link-light;
     }
 
@@ -97,7 +101,6 @@
     }
 
     .collapsed {
-
     }
   }
 }
